server: task: Fix segfault when parsing fails.
When the command_line is empty g_shell_parse_argv will fail, but is not
filling the error variable. So we will get an segmentation fault when we
try to print an error message. To avoid this, we only print the additional
error info, when we have verified that it is valid. Also we return with
-EINVAL, when the command line argument is not valid.

Change-Id: I2551f257e5056f0e4fdf5ae987c37480da02a1cd
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/349
Reviewed-by: Julian Scheel <julian.scheel@avionic-design.de>