server: vlc: Implement configurable buffer duration.
For the gstreamer backend we have a configurable buffer duration. For the vlc
backend this was not implemented because, we could set the buffer duration
via the uri. Now that we have switched from gstreamer to vlc for medcom-wide,
we have the problem, that the McAdmin database does not provide these
information, but the configuration file still does.

Change-Id: I891140a2e225617ef3b446e1e6fcb8cb129a8ebb
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/352
Reviewed-by: Julian Scheel <julian.scheel@avionic-design.de>