server: Pass audio-backend to sound-manager.
Because on medcom-wide Gen2 the handset can be a usb-handset, which provides
a separate soundcard, we need to do the playback on this soundcard, when we
are in audio-state handset. Therefor we need to know that we are in such an
audio-state, so we need a way to query the audio-state and therefor we need
a reference to the audio-backend.

Change-Id: Ib34f7476f17380d41f678322f108acb1f17f9ec9
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/385
Reviewed-by: Julian Scheel <julian.scheel@avionic-design.de>