server: sound-manager-vlc: Change output device for handset audio state.
On medcom-wide we can have a usb-handset, which is actually a different sound-
card. So if we are in a audio-state were we need to output to the handset we
need to ensure, that we are using the right sound-card.

We could enumerate over all cards, to figure out which card we could use. But
then we still don't know which card is the right one. This could only be set
via configuration. But then needs to be done manually. So we decided to create
a new alsa plug which can be used. So the os needs to select the right card
for us.

Change-Id: Ib5f9d9e51fc3da8b9fa3532fcb2e9e36af57d3ea
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/386
Reviewed-by: Julian Scheel <julian.scheel@avionic-design.de>