backlight: medatom: Use only one backend at a time.
The medatom backend uses DPMS to enable/disable the backlight and either
sysfs or I2C to control the brightness. Because we are going to disable DPMS,
we need to change the backend initialisation. This requires some code
restructuring, so we can use only sysfs, I2C or DPMS. Therefore we need to
complete the existing implementations and change the internal structure, so
we are able to run with one backend exclusively. Finally we change the
probing, to use the best fitting backend. We probe first sysfs, then I2C and
only as fallback DPMS.

Change-Id: I44e14ddb79b0f4b62ad625dd6bed6a91fb58d4d7
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/812
Reviewed-by: Julian Scheel <julian@jusst.de>