remote-control: Restart only on usb-handset add detection
In real life it only makes sense to restart remote-control if a device is
added. On removal a device usually is no longer present and can thus be
ignored. Additionally a restart in progress from removal could lead to
missing subsequent add events if a headset is changed fast enough.

Change-Id: I6ef2cbc2df74081351b895e9a07b322eb0948edb
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/887
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>