bin/remote-control: prevent possible circular call
The helper functions in javascript.c introduced in d8fa62c5 accidentally
created a circular call possibility for java_script_make_string() and
javascript_set_exception_text(). In case the former would fail, the
latter was called which used the former again to create a string for an
exception. If string creation keeps failing, the system blows up.

Change-Id: I63499cac6330bae8376bc95447de8b76977a2512
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1641
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>