bin/remote-control: Add a JS API to control the outputs
This API allow controlling outputs, currently using sysfs or a similar
filesystem based API. This implementation allow the user to set/get
the output value, as well as running a predefined sequences of set.
The later allow to set an output for a given time without having to
use any timer in javascript.

The sysfs backend allow using a predefined path or a udev match.
Hotplug is currently missing, however as the device lookup is done
on page load this shouldn't be a show stopper.

Change-Id: I2998e36ba61802d5b6b15427c567d180a529672e
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1651
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>