javascript: Rework the outputs config to allow multiple config files
Currently the outputs are all defined in one group with a key listing
all the existing outputs. This has the problem that we can't combine
several config files as only the last name list would be taken.

Rework this to use one group per output, the groups must be named
"output $NAME" and contain at least one "type" key with the proper type.

Change-Id: I7989e9fefa6278a1eb32082d0b0d6784a00b7320
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1675
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>