bin/remote-control: Fix page reload on failure
If the network connectivity is not fully setup when remote-control starts the
start page might not be loaded successfully. To deal with this an automatic
reload is scheduled. With recent webkit versions the reload command might be a
no-op depending on the type of the previous error though. So to be sure that
the reload will take place schedule an explicit new load_uri instead of a
reload.

Change-Id: Ibff91dce35203a28137d02aef233eaed2256eb3e
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/1807
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>