bin/remote-control: Add watchdog autostart option
Introduce a new configuration option to remote-control.conf, which allows
setting a default timeout which is used to autostart the javascript watchdog
as soon as the javascript context is loaded.
This allows catching freezes which happen even before the html page was
properly loaded. Examples for this are server failures where only parts of the
page content are transferred and webkit aborts parsing somewhere inbetween.

Change-Id: Iddae24800b6d0c40c2561d9fa3557f86e2d4a42a
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/1812
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>