server: Add application watchdog
This implements a global application watchdog as part of the remote-control
server core. The code is virtually the same as we use in the
javascript-app-watchdog so far. The javascript-app-watchdog is meant to be
ported to use this module instead of it's own implementation now. The benefit
of this restructuring is that only one global watchdog instance exists, no
matter how many webkit contexts are alive. From a user perspective nothing
changes as the user assumption has always been that there is a single watchdog
behind the static javascript module.

Change-Id: I939b6cb6090c5d24455d71a0a1f3ad89c41bf826
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/1843
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>