server: app-watchdog: use dbus-watchdog service
As the app-watchdog nowadays has not only to take care of the javascript
application but also of webkit itself it turned out to be problematic having
the watchdog in the same process as webkit. Especially within the same
mainloop.  The medcom devices already include a dedicated watchdog service
called dbus-watchdog, which this patch uses as backend for the application
watchdog.

Change-Id: I484a1c4ffc9108e62aa560fd269fe68304e90deb
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/1977
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>