server: audio-manager: vlc: Use configured values for volume
It is now set to the same value as media player, since it also has an effect
there. To use the configured value for volume, we also need the config file
object in the create function. The log messages in the media player are changed
to differentiate them from the ones in the audio manager.

Change-Id: I5017666162ef5818d9f6fc3b1c3e3385d03874d3
Signed-off-by: Dirk Leber <dirk.leber@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/2721
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>