server: Add support for es changed callback for vlc 2.0.7
Since vlc only supports es notifications for versions greater that 3 and we
already depend on this feature, we need to add it for vlc 2.0.7. We use the
position changed callback to check the current audio and spu pids. Video is
not monitored, since we don't expect a video stream selection.

Change-Id: Iacde329867808bd2b1642d1e265eb95f6fa53bc0
Signed-off-by: Dirk Leber <dirk.leber@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/2942
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>