javascript: Use same error reporting for exceptions in callbacks
Always use the same text for exceptions in callbacks by giving a hint that they
should be cleared on unload. Prevent printing the exception since it will flood
the log but always print a warning if the handler fails, except for EFAULT since
this is reported in the handler itself.

Change-Id: Icf226bfdc04b1d304d5f6d95de55ce39227ee96a
Signed-off-by: Dirk Leber <dirk.leber@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4020
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>