server: vlc: Set state to stopped if not running
We have to set the state to stopped if the stream is not running because
the callback function that will set this state will never be called.
Just to be safe we still call the vlc stop function even if it should
not have an effect (but it also doesn't hurt).

Change-Id: I42d049f57c5ef9beb8dae9c3e5e68fe44ecff69c
Signed-off-by: Dirk Leber <dirk.leber@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4641
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>