server, bin/remote-control, js: replace g_error
The logging function g_error() has the side effect of producing a stack
dump and aborting the entire process. As the code design does not look
like that was intentional and any cleanup functions will not be called
with g_error() replace it with g_critical() to make a graceful exit.
Prevents large memleaks.

Note there is one place where a g_error() call is left intentionally as
comments indicate a remote-control restart is wanted. Additionally
another called is erased entirely as the check for a failing g_new0() is
pointless.

Change-Id: Icadd4e5a1a2802f53187cc2702e5a80c7b40e156
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5667
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>