Revert "bin/remote-control: let startup signal time out"
This reverts commit f7aaa67f46b5c369ffa802824f6c2910afdec721.
It does not make sense to timeout waiting for the startup_cond, since
g_cond_wait_until can only return if the mutex is unlocked and that
is always after initialization or in case of an error.
This patch also breaks debugging, since initialization can take more time. The
result is that after successful initialization, the functions fails due to the
timeout.
But there was a case where startup took endless and that can only occur if we
have an fast return due to an error. In that case the startup_cond never gets
signalled. I will fix this in the next patch.

Change-Id: I431f5d8b6fcc981c85271dd30f0a802623413555
Signed-off-by: Dirk Leber <dirk.leber@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5849
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>