lldp-monitor: Introduce info function
Provide a function for the lldp-monitor that collects all lldp
informations in a hash list so that we don't need to parse the binary
package format. This information is used to create a javascript object
accessible via AvionicDesign.LLDP.info. Note that this is only
implemented for lldpctl, since the old implementation is somehow
deprecated.

Change-Id: Ifa461787fb4ab65f1efb0b62262361f61d2335dc
Signed-off-by: Dirk Leber <dirk.leber@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5999
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>