src/lib: add back net-udp in generic form
The network layer for UDP traffic went down the drain when the RPC
interface was killed, but there still is a need for it. Rebuild it,
loosely inspired by the old one. Instead of one configurable and one
hard-coded channel a generic number of channels is supported, with
access through the JS API in mind. A small test perfoming a single
packet exchange is included.

Change-Id: I8781dd8e87385146cf54273c524984afbc2d0fc4
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6651
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>