src/server/event-manager: split RPC part from event-manager
Pull out the RPC part from the event-manager and put it into irq.c,
where related RPC stuff already exists. To achieve this, a callback for
event-manager events is added and irq.c gets its own static data
structure. A cleanup function to free this data structure is added to
the API for now.

Change-Id: Ia9bc3caaa9e1988e35f3d7472418908b6e5e055f
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7672
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>