src/server/net-udp: use channel based callbacks
Instead of having a single callback for any packet-received events on
any channel, use callbacks on a per channel basis. This avoids invoking
callbacks for channels which were not created by the callback owner. Use
the channel data pointer as an owner reference while still supplying it
as an argument to the callback to conviently read available data right
there. Note that support for multiple callbacks has not yet been
implemented but is planned in the near future.

Change-Id: I9485e9a7ca7f7cd7bbe96187e5d6e28593f0d727
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7709
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>