src/server/utils: disable gdk_window_clear for Gtk+ 3.20
The gdk_window_clear function already ceased to exist in Gtk+ 2.91 and
had a rather ugly replacement. Since those now also became deprecated
with Gtk+ 3.20 and it is doubtful whether gdk_window_clear is really
needed for the use cases in remote-control, disable the entire function
to an empty body.

Change-Id: Iffbd93f75ba6fe8968a1dcc2660465892d7baee3
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7715
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>