src/server/net-udp: relax polling thread mutex locks
The file descriptors polling thread locked the mutex at the beginning of
its loop. When file descriptors are set but have no events, the thread
unlocked the mutex and returned to the beginning of its loop. This
however left no breathing space for others to get the mutex locked. Make
the thread sleep a bit and reduce the polling timeout to get the same
polling rate and still give others a chance for the mutex.

Change-Id: I415527596250f26af90c6ea51f9d704ce38f5f3b
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7773
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>