src/server: split rpc and remote-control data memory slab
Instead of having a single piece of memory allocated for both RPC server
and remote-control private data, use distinct allocations to make
remote-control work indepedently from RPC. As the latter still needs a
way to get a reference for the former's data, put a pointer into the
server's space for remote-control data. This consequently adds a level
of pointer dereference for any remote-control data calls from RPC, which
is covered by a tiny macro.

Change-Id: Id4fe95588594ea92e2e6fb243fa3347c3f0bb2d4
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7775
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>