src/core: use wrapper for gdk_window_clear
Instead of a GTK version check and optional definition of
gdk_window_clear(), make a wrapper function in utils.c with its
implementation depending on the GTK version. While the wrapper falls a
bit out of the gdk_ naming scheme, this has the huge benefit of dropping
the gtk header inclusion in the core remote-control.h header. That way
extensions building against libremote-control don't need to link against
gtk themselves.

Change-Id: I99fb0b0f07e89bbcb25d44878f57052d8de8ee5e
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7809
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>