src/core/smartcard-nxp: add I2C write mode
Add support for writing I2C cards. Note that standard instead of
extendend mode is used for single byte offsets as tiny cards are the
primary use case. Since extended mode uses two instead of one byte for
the offset, this might probably not work large cards. However, a use
case for those remains yet to be defined, so stick with it for now.

Change-Id: I25290360b7972d5bda3345c247e0fe83df5eccab
Signed-off-by: Bert van Hall <bert.vanhall@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7916
Reviewed-by: Meike Vocke <meike.vocke@avionic-design.de>